Allow pip to intall source packages if there are no binaries
Remove the `--only-binary=:all:` flag from the `pip install`
command. This option prevents pip from installing _any_ packages
from binary wheels, even those packages that only have source
distributions or that lack binary wheels for the target platform.

Note that with this change there _is_ a risk that when hydrating
for a foreign architecture (e.g. when hydrating for aarch64
deployment on an i86_64 system) that pip may build wheels for the
wrong architecture if compilers are not properly configured for
cross-compilation.

However, with this change it is at _least_ possible to build
packages from source when necessary and when the host and target
archs match.

For concreteness, we encountered the bug that this patch fixes
when we were trying to hydrate a package that depends on
[logbook 1.7.0](https://pypi.org/project/Logbook/1.7.0.post0/#files)
for aarch64, because that package has no aarch64 wheels at all.